videoDrawer script
Still needs to be thoroughly gone through, seems there is an issue where too many frames are generated resulting in silent animation at the end of some videos. Revision of blink iteration and possibly a more appropriate cycling method of the poses. Look into how the scheduler appoints a pose